+++ CHANGELOG.md
@@ -10,6 +10,9 @@
  - Require units for duration configuration (#211)
  - Add support for multiple server URLs with failover (#233)
  - Add support for mixing OpenTracing spans with native transactions/spans (#235)
+ - Drop SetHTTPResponseHeadersSent and SetHTTPResponseFinished methods from Context (#238)
+ - Stop setting custom context (gin.handler) in apmgin (#238)
+ - Set response context in errors reported by web modules (#238)
 
 ## [v0.5.0](https://github.com/elastic/apm-agent-go/releases/tag/v0.5.0)